requirements.txt
# Application dependencies
python-dotenv==0.17.1

# Database dependencies
psycopg2==2.8.2

# Server dependencies
django==3.2.5
gunicorn==20.1.0
whitenoise==5.1.0
django-cors-headers==3.7.0
django-referrer-policy==1.0.0

# API dependencies
djangorestframework==3.12.4